requirements.txt
pillow>=10.0.0
pytesseract>=0.3.10
reportlab>=4.0.0
python-docx>=0.8.11
opencv-python>=4.8.0
numpy>=1.24.0
tkinter-tooltip>=2.1.0
tkinter-dnd>=0.3.0